Refactor setup in setup folder
