+++ setup/check_local_modules.py
