+++ setup/setup_windows.py
